requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.2.1
django-crispy-forms==2.4
django-model-utils==5.0.0
pillow==11.2.1
python-dotenv==1.1.0
sqlparse==0.5.3
